Refactor: Split Neuron.ts into focused modules (#1599)
Extract Neuron.ts (1,339 lines) into five focused modules in src/neuron/:
- NeuronActivation.ts: Activation function management and computation
- NeuronPropagation.ts: Backpropagation and error distribution
- NeuronRecord.ts: Error recording for structural discovery
- NeuronTopology.ts: Topology fixing, mutation, and connectivity
- NeuronSerialization.ts: JSON import/export

Neuron.ts is now a ~290-line facade that delegates to these modules,
following the same pattern used for Creature in src/creature/.
